Assign port numbers in server config file; divide Python shared library into multiple files
