+++ goss-testing/dat/goss-servers.cfg
@@ -1,38 +1,35 @@
 # This configuration file defines the Goss test suite endpoints that will be started on NCNs.
 #
 # Designated goss-servers port range: 8994-9008
-# Ports will be assigned automatically to the endpoints, starting with the start_port, going no higher than end_port.
-# For test suites which run on multiple NCN types, it may not have the same port number on each type (however, it will
-# have the same port number on all NCNs of a given type).
 
 # In this file, lines beginning with # and lines with only whitespace are ignored.
 
-start_port  8994
-end_port    9008
+#port      Suite file                                              NCN type(s) where it will be run
+8994 	   ncn-preflight-tests.yaml                                master storage worker
 
-#           Suite file                                              NCN type(s) where it will be run
-suite       ncn-healthcheck-master.yaml                             master
-suite       ncn-healthcheck-master-single.yaml                      master
-suite       ncn-healthcheck-storage.yaml                            storage
-suite       ncn-healthcheck-worker.yaml                             worker
-suite       ncn-healthcheck-worker-single.yaml                      worker
+8995       ncn-smoke-tests.yaml                                    master storage worker
 
-suite       ncn-afterpitreboot-healthcheck-master.yaml              master
-suite       ncn-afterpitreboot-healthcheck-storage.yaml             storage
-suite       ncn-afterpitreboot-healthcheck-worker.yaml              worker
-suite       ncn-afterpitreboot-healthcheck-worker-single.yaml       worker
+8996       ncn-spire-healthchecks.yaml                             master storage worker
 
-suite       ncn-kubernetes-tests-master.yaml                        master
-suite       ncn-kubernetes-tests-master-single.yaml                 master
-suite       ncn-kubernetes-tests-worker.yaml                        worker
+8997       ncn-healthcheck-master.yaml                             master
+8997       ncn-healthcheck-storage.yaml                            storage
+8997       ncn-healthcheck-worker.yaml                             worker
 
-suite       ncn-afterpitreboot-kubernetes-tests-master-single.yaml  master
-suite       ncn-afterpitreboot-kubernetes-tests-worker-single.yaml  worker
+8998       ncn-healthcheck-master-single.yaml                      master
+8998       ncn-healthcheck-worker-single.yaml                      worker
 
-suite       ncn-preflight-tests.yaml                                master storage worker
+8999       ncn-afterpitreboot-healthcheck-master.yaml              master
+8999       ncn-afterpitreboot-healthcheck-storage.yaml             storage
+8999       ncn-afterpitreboot-healthcheck-worker.yaml              worker
 
-suite       ncn-storage-tests.yaml                                  storage
+9000       ncn-afterpitreboot-healthcheck-worker-single.yaml       worker
 
-suite       ncn-smoke-tests.yaml                                    master storage worker
+9001       ncn-kubernetes-tests-master.yaml                        master
+9001       ncn-kubernetes-tests-worker.yaml                        worker
 
-suite       ncn-spire-healthchecks.yaml                             master storage worker
+9002       ncn-kubernetes-tests-master-single.yaml                 master
+
+9003       ncn-afterpitreboot-kubernetes-tests-master-single.yaml  master
+9003       ncn-afterpitreboot-kubernetes-tests-worker-single.yaml  worker
+
+9004       ncn-storage-tests.yaml                                  storage
